[Refactor] CustomTabBar 개별 View로 리팩토링
